django-backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-decouple==3.8
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
django-filter==24.2
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
kaggle==1.5.16
beautifulsoup4==4.12.3
lxml==5.1.0
